requirements.txt
# Core numeric
numpy==1.26.4
pillow==10.4.0

# Data / IO
pandas==2.2.2
pyarrow==16.1.0
h5py==3.11.0

# Sci stack
scipy==1.13.1
scikit-learn==1.5.1

# CV
opencv-python-headless==4.10.0.84
albumentations==1.4.13
imageio==2.34.2
imageio-ffmpeg==0.5.1

# Metrics
lpips==0.1.4
pytorch-fid==0.3.0

# Utils / logging
pyyaml==6.0.2
tensorboard==2.17.0
wandb==0.17.7
tqdm==4.66.5
einops==0.8.0
timm==1.0.8
requests==2.32.3

# Dataset tooling (nuScenes)
nuscenes-devkit==1.1.11

# Visualization (필요하면)
# matplotlib>=3.7.0
# seaborn>=0.12.0
# moviepy>=1.0.3
# waymo-open-dataset-tf-2-11-0>=1.6.0  # For Waymo dataset (optional)

# Optional: Advanced Features
# torchmetrics>=1.0.0
# pytorch-lightning>=2.0.0  # For easier training loops
# hydra-core>=1.3.0  # For advanced configuration management

# Optional: Distributed Training
# deepspeed>=0.10.0
# fairscale>=0.4.13

# Optional: Model Compression
# torch-pruning>=1.2.0
# neural-compressor>=2.3.0
